requirements.txt
yfinance
psycopg2-binary
python-dotenv
